requirements.txt
Django~=5.1.7
django-extensions~=3.2.3
djangorestframework~=3.16.0
django-cors-headers~=4.7.0

psycopg2
geojson

pycodestyle

gunicorn
